requirements.txt
streamlit
tensorflow
google-generativeai
pandas
numpy
joblib
scikit-learn
